refactor: Move logging to core
