+++ packages/ref-celery/src/cmip_ref_celery/app.py
@@ -12,6 +12,7 @@ from loguru import logger
 from rich.pretty import pretty_repr
 
 from cmip_ref.config import Config
+from cmip_ref_core.logging import capture_logging
 
 os.environ.setdefault("CELERY_CONFIG_MODULE", "cmip_ref_celery.celeryconf.dev")
 
@@ -33,8 +34,6 @@ def create_celery_app(name: str) -> Celery:
 @setup_logging.connect
 def setup_logging_handler(loglevel: int, **kwargs: Any) -> None:  # pragma: no cover
     """Set up logging for the Celery worker using the celery signal"""
-    from cmip_ref.cli._logging import capture_logging
-
     capture_logging()
 
     # Include process name in celery logs
