--- packages/ref/src/cmip_ref/cli/_logging.py
@@ -1,45 +1,0 @@
-import inspect
-import logging
-
-import pooch
-from loguru import logger
-
-
-class _InterceptHandler(logging.Handler):
-    def emit(self, record: logging.LogRecord) -> None:
-        # Get corresponding Loguru level if it exists.
-        level: str | int
-        try:
-            level = logger.level(record.levelname).name
-        except ValueError:  # pragma: no cover
-            level = record.levelno
-
-        # Find caller from where originated the logged message.
-        frame, depth = inspect.currentframe(), 0
-        while frame and (depth == 0 or frame.f_code.co_filename == logging.__file__):
-            frame = frame.f_back
-            depth += 1
-
-        logger.opt(depth=depth, exception=record.exc_info).log(level, record.getMessage())
-
-
-def capture_logging() -> None:
-    """
-    Capture logging from the standard library and redirect it to Loguru
-
-    Note that this replaces the root logger, so any other handlers attached to it will be removed.
-    """
-    # Pooch adds a handler to its own logger which circumvents the REF logger
-    pooch.get_logger().handlers.clear()
-    pooch.get_logger().addHandler(_InterceptHandler())
-
-    logging.basicConfig(handlers=[_InterceptHandler()], level=0, force=True)
-
-    # Disable some overly verbose logs
-    logger.disable("matplotlib.colorbar")
-    logger.disable("matplotlib.ticker")
-    logger.disable("matplotlib.font_manager")
-    logger.disable("pyproj.transformer")
-
-
-__all__ = ["capture_logging", "logger"]
